+++ app/scripts/controllers/permissions/enums.js
@@ -70,6 +70,7 @@ export const SAFE_METHODS = [
   'eth_signTypedData',
   'eth_signTypedData_v1',
   'eth_signTypedData_v3',
+  'eth_signTypedData_v4',
   'eth_submitHashrate',
   'eth_submitWork',
   'eth_syncing',