+++ old-ui/app/components/editable-label.js
@@ -52,9 +52,9 @@ EditableLabel.prototype.saveIfEnter = function (event) {
 
 EditableLabel.prototype.saveText = function () {
   // eslint-disable-next-line react/no-find-dom-node
-  var container = findDOMNode(this)
-  var text = container.querySelector('.editable-label input').value
-  var truncatedText = text.substring(0, 20)
+  const container = findDOMNode(this)
+  const text = container.querySelector('.editable-label input').value
+  const truncatedText = text.substring(0, 20)
   this.props.saveText(truncatedText)
   this.setState({ isEditingLabel: false, textLabel: truncatedText })
 }