+++ old-ui/css.js
@@ -3,7 +3,7 @@ const path = require('path')
 
 module.exports = bundleCss
 
-var cssFiles = {
+const cssFiles = {
   'fonts.css': fs.readFileSync(path.join(__dirname, '/app/css/fonts.css'), 'utf8'),
   'reset.css': fs.readFileSync(path.join(__dirname, '/app/css/reset.css'), 'utf8'),
   'lib.css': fs.readFileSync(path.join(__dirname, '/app/css/lib.css'), 'utf8'),
@@ -21,9 +21,9 @@ var cssFiles = {
 }
 
 function bundleCss () {
-  var cssBundle = Object.keys(cssFiles).reduce(function (bundle, fileName) {
-    var fileContent = cssFiles[fileName]
-    var output = String()
+  const cssBundle = Object.keys(cssFiles).reduce(function (bundle, fileName) {
+    const fileContent = cssFiles[fileName]
+    let output = String()
 
     output += '/*========== ' + fileName + ' ==========*/\n\n'
     output += fileContent